+++ front/src/components/contextData/vulnerability/VulnerabilityMainContextDataScore.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { computed } from "vue"
-import { getVulnerabilityColor } from "@/utils/vulnerability"
+import { getVulnerabilityTextColor } from "@/utils/color"
 
 interface VulnerabilityMainScoreProps {
   score: number
@@ -10,23 +10,7 @@ interface VulnerabilityMainScoreProps {
 
 const props = defineProps<VulnerabilityMainScoreProps>()
 
-const textColor = computed(() => {
-  if (!props.score) return "text-gray-300"
-  const colorCode = getVulnerabilityColor(props.score)
-  const colorMap: Record<string, string> = {
-    "#4474b5": "text-blue-600",
-    "#75add1": "text-blue-400",
-    "#aad9e9": "text-blue-200",
-    "#5aaf7b": "text-green-600",
-    "#9cbf4e": "text-green-500",
-    "#d7e360": "text-green-300",
-    "#fdae60": "text-orange-300",
-    "#f56c43": "text-orange-500",
-    "#d73026": "text-red-600",
-    grey: "text-gray-500"
-  }
-  return colorMap[colorCode] || "text-gray-500"
-})
+const textColor = computed(() => getVulnerabilityTextColor(props.score))
 </script>
 
 <template>
